Moves Tab into TabContainer directory, cleans up styles.
