+++ react/src/components/organisms/TabContainer/TabContainer.stories.js
@@ -1,12 +1,10 @@
 import React from 'react';
-
 import { storiesOf } from '@storybook/react';
 import { withInfo } from '@storybook/addon-info';
-import { withKnobs, object, select, text, boolean } from '@storybook/addon-knobs/react';
-import { action } from '@storybook/addon-actions';
+import { withKnobs, boolean } from '@storybook/addon-knobs/react';
 
 import TabContainer from '.';
-import Tab from '../../molecules/Tab';
+import Tab from './tab';
 
 storiesOf('organisms', module).addDecorator(withKnobs)
   .add(
