+++ react/src/components/organisms/TabContainer/context.js
@@ -1,7 +1,6 @@
-// Make sure the shape of the default value passed to
-// createContext matches the shape that the consumers expect!
 import React from 'react';
 
+// Used by @Organisms/TabContainer and Tab to share data.
 export const TabContext = React.createContext({
   activeTab: null,
   activeContent: null,
