+++ react/src/components/organisms/TabContainer/style.scss
@@ -11,19 +11,26 @@
     -webkit-overflow-scrolling: touch;
     -ms-overflow-style: -ms-autohiding-scrollbar;
     .ma__tab-title, %ma__tab-title {
-      border: 1px solid $c-gray-light;
+      border-color: $c-gray-light;
+      outline-color: $c-gray-light;
       cursor: pointer;
       padding: 1.5rem 1.5rem 0.8rem 1.5rem;
       font-size: 1.25rem;
-      color: inherit;
-
+    }
+    .ma__tab-title:focus, .ma__tab-title--active:focus {
+      outline: none !important;
+      border-color: $c-gray-light !important;
+      border-top-color: $c-primary !important;
+      outline-color: $c-gray-light !important;
+      border-left-color: $c-gray-light !important;
+      border-bottom: none;
     }
     .ma__tab-title--active {
       @extend %ma__tab-title;
-      border-bottom: none;
-      border-top: 5px solid $c-primary;
       color: $c-primary;
+      border-bottom: none;
       font-weight: 700;
+      border-top: 5px solid $c-primary;
     }
     .ma__tab-title ~ .ma__tab-title {
       border-left: 1px solid $c-gray-light;
@@ -62,9 +69,18 @@
         @extend %ma__tab-title;
         border: none;
       }
+      .ma__tab-title:focus, .ma__tab-title--active:focus {
+        outline: none;
+        border-bottom-width: 5px !important;
+        border-bottom-style: solid !important;
+        border-bottom-color: $c-primary !important;
+      }
       .ma__tab-title--active {
         @extend %ma__tab-title;
-        border-bottom: 5px solid $c-primary;
+        color: $c-primary;
+        border-bottom-width: 5px !important;
+        border-bottom-style: solid !important;
+        border-bottom-color: $c-primary !important;
       }
       .ma__tab-title ~ .ma__tab-title {
         border-left: none;
